USMAV3/USMAProject/requirements.txt
Django==4.2
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.13.1
djoser==2.2.0
drf-yasg==1.21.7
Pillow==9.5.0



